+++ R/exclude_nca.R
@@ -9,10 +9,17 @@
 #'   half-life (uses \code{PKNCA.options("min.hl.r.squared")} if not
 #'   provided).
 #' @examples
-#' my_conc <- PKNCAconc(data.frame(conc=1.1^(3:0), time=0:3, subject=1), conc~time|subject)
-#' my_data <- PKNCAdata(my_conc, intervals=data.frame(start=0, end=Inf, aucinf.obs=TRUE, aucpext.obs=TRUE))
+#' my_conc <- PKNCAconc(data.frame(conc=1.1^(3:0),
+#'                                 time=0:3,
+#'                                 subject=1),
+#'                      conc~time|subject)
+#' my_data <- PKNCAdata(my_conc,
+#'                      intervals=data.frame(start=0, end=Inf,
+#'                                           aucinf.obs=TRUE,
+#'                                           aucpext.obs=TRUE))
 #' my_result <- pk.nca(my_data)
-#' my_result_excluded <- exclude(my_result, FUN=exclude_nca_max.aucinf.pext())
+#' my_result_excluded <- exclude(my_result,
+#'                               FUN=exclude_nca_max.aucinf.pext())
 #' as.data.frame(my_result_excluded)
 #' @name exclude_nca
 NULL